Refactored Role addition to the User

+++ src/main/java/com/jarad/postly/service/ProfileServiceImpl.java
@@ -4,15 +4,19 @@ import com.jarad.postly.entity.Comment;
 import com.jarad.postly.entity.Follower;
 import com.jarad.postly.entity.Post;
 import com.jarad.postly.entity.Profile;
+import com.jarad.postly.entity.Role;
 import com.jarad.postly.entity.User;
 import com.jarad.postly.repository.CommentRepository;
 import com.jarad.postly.repository.FollowerRepository;
 import com.jarad.postly.repository.PostRepository;
 import com.jarad.postly.repository.ProfileRepository;
+import com.jarad.postly.repository.RoleRepository;
 import com.jarad.postly.repository.UserRepository;
 import com.jarad.postly.util.dto.ProfileDto;
+import com.jarad.postly.util.enums.SecurityRole;
 import com.jarad.postly.util.exception.ProfileForUserAlreadyExistException;
 import com.jarad.postly.util.exception.ProfileNotFoundException;
+import com.jarad.postly.util.exception.RoleNotFoundException;
 import com.jarad.postly.util.exception.UserNotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -37,9 +41,9 @@ import static java.util.stream.Collectors.toSet;
 @Transactional(readOnly = true)
 @Slf4j
 public class ProfileServiceImpl implements ProfileService {
-
     public static final int PAGE_SIZE = 10;
     public static final String CREATION_DATE = "creationDate";
+    private final RoleRepository roleRepository;
     private final ProfileRepository profileRepository;
     private final FollowerRepository followerRepository;
     private final PostRepository postRepository;
@@ -47,12 +51,14 @@ public class ProfileServiceImpl implements ProfileService {
     private final CommentRepository commentRepository;
 
     @Autowired
-    public ProfileServiceImpl(ProfileRepository profileRepository, FollowerRepository followerRepository, PostRepository postRepository, UserRepository userRepository, CommentRepository commentRepository) {
+    public ProfileServiceImpl(ProfileRepository profileRepository, FollowerRepository followerRepository, PostRepository postRepository, UserRepository userRepository, CommentRepository commentRepository,
+                              RoleRepository roleRepository) {
         this.profileRepository = profileRepository;
         this.followerRepository = followerRepository;
         this.postRepository = postRepository;
         this.userRepository = userRepository;
         this.commentRepository = commentRepository;
+        this.roleRepository = roleRepository;
     }
 
     @Override
@@ -125,6 +131,8 @@ public class ProfileServiceImpl implements ProfileService {
 
         User user = optionalUser.get();
         user.setActiveProfile(true);
+        user.addRole(getOrCreateRole(SecurityRole.ROLE_PROFILE_ACTIVE));
+
         Profile profile = Profile.builder()
                 .username(profileDto.getUsername())
                 .creationDate(Instant.now())
@@ -170,11 +178,21 @@ public class ProfileServiceImpl implements ProfileService {
             throw new ProfileNotFoundException(message);
         }
 
-        User user = optionalUser.get();
-        user.setActiveProfile(false);
+        Optional<Role> optionalRole = roleRepository.findByName(SecurityRole.ROLE_PROFILE_ACTIVE.toString());
+        if (optionalRole.isEmpty()) {
+            String message = MessageFormat.format("Role for user with name {0} doesn''t exist", SecurityRole.ROLE_PROFILE_ACTIVE);
+            log.warn(message);
+            throw new RoleNotFoundException(message);
+        }
 
+        User user = optionalUser.get();
+        Role role = optionalRole.get();
+        
         profileRepository.deleteByUserAndId(user, profileId);
 
+        user.setActiveProfile(false);
+        user.removeRole(getOrCreateRole(SecurityRole.ROLE_PROFILE_ACTIVE));
+        userRepository.save(user);
         log.info("Profile with ID {} for user with ID {} has been deleted", profileId, userId);
     }
 
@@ -202,5 +220,27 @@ public class ProfileServiceImpl implements ProfileService {
         }
 
         return Collections.emptySet();
+    }
+
+    /**
+     * Helper method that checks if ROLE_USER already exists in Database. If not creates it
+     *
+     * @return ROLE_USER from table 'roles'
+     */
+    @Transactional
+    public Role getOrCreateRole(SecurityRole securityRoleName) {
+        log.info("Retrieving the {} role", securityRoleName);
+        Optional<Role> optionalRole = roleRepository.findByName(securityRoleName.toString());
+
+        if (optionalRole.isEmpty()) {
+            log.info("Creating and saving the {} role", securityRoleName);
+            Role roleUser = Role.builder()
+                    .name(securityRoleName.toString())
+                    .build();
+            roleRepository.save(roleUser);
+            return roleUser;
+        }
+
+        return optionalRole.get();
     }
 }
